Add Gain RAMSES adaptation: CSV sink reader and HaloQ struct fix
- NewDD: add rd_sink_csv.c to read 23-col RAMSES sink CSV (replaces
  binary rd_sink.c usage in newdd.c). Disable QUADHILBERT/READ_SINK/
  NCHEM/NDUST flags. Update rd_part.c for the simpler particle layout
  this RAMSES build emits (no mass0/chem/birth_d/partp).
- NewGalFinder: HaloQ.{mass,mstar,mgas,mdm,msink} float -> double to
  match opfof's on-disk layout (was a 96 vs 120 byte mismatch causing
  silent gfind hangs after the first halo). Drop unused -DREAD_SINK/
  -DNCHEM=3, lower default heap (NMEG 90000->8000, master 80->16 GB).
- opFoF: switch compiler to mpiicx/mpiifx, drop QUADHILBERT/READ_SINK/
  NCHEM/NDUST, restore boolean enum in fof.h for the FoF build path.

+++ NewGalFinder/tree.h
@@ -55,7 +55,7 @@ typedef struct Box{
 typedef struct HaloQ{
     size_t np,npstar,npgas,npdm,npsink;
     POSTYPE x,y,z;
-    float mass, mstar,mgas,mdm,msink;
+    double mass, mstar,mgas,mdm,msink;
     float vx,vy,vz;
 }HaloQ;
